Use LocalizedPrice for discount offers.
