+++ src/payment/data/service.js
@@ -8,7 +8,6 @@ import handleBasketApiError from '../utils/handleBasketApiError';
 import { camelCaseObject } from '../../common/utils';
 import { ORDER_TYPES } from './constants';
 
-
 let config = {
   ACCOUNTS_API_BASE_URL: null,
   ECOMMERCE_BASE_URL: null,
@@ -79,12 +78,17 @@ export function transformResults(data) {
   const lastProduct = results.products && results.products[results.products.length - 1];
   results.orderType = getOrderType(lastProduct && lastProduct.productType);
   results.orderTotal = Number(results.orderTotal);
-  results.summaryDiscounts = results.summaryDiscounts !== null ?
-    Number(results.summaryDiscounts) : null;
+  results.summaryDiscounts =
+    results.summaryDiscounts !== null ? Number(results.summaryDiscounts) : null;
   results.summaryPrice = Number(results.summaryPrice);
 
   if (results.offers != null) {
-    results.offers = results.offers.filter(({ provider }) => provider !== null);
+    results.offers = results.offers
+      .filter(({ provider }) => provider !== null)
+      .map(offer => ({
+        ...offer,
+        benefitValue: Number(offer.benefitValue),
+      }));
   }
 
   return results;
@@ -127,4 +131,3 @@ export async function deleteCoupon(id) {
     .catch(handleBasketApiError);
   return transformResults(data);
 }
-
