+++ src/payment/order-summary/index.jsx
@@ -1,0 +1,3 @@
+import OrderSummary from './OrderSummary';
+
+export default OrderSummary;
